+++ src/index.js
@@ -1,85 +1,102 @@
-let fs = require('fs');
-let rollup = require('rollup').rollup;
+'use strict';
+let rollup = require('rollup');
 let debounce = require('debounce');
 let dependencies = new Map();
 let changedParents = new Set();
 let WAIT = 25;
 
-// Workaround for nodejS 0.12 -> don't use ES1015 for this function
-const touchParents = debounce ( function () {
+const touchParents = debounce(function () {
     let now = new Date();
     for (let idx = 0, lst = changedParents.values(); idx < lst.length; idx += 1) {
-        fs.utimes(lst[idx], now, now);
+        fs.utimes(parent, now, now);
     }
     changedParents.clear();
-}, WAIT);
+}, WAIT)
 
-function createPreprocessor (args, config = {}, logger, helper) {
+// @param args {Object} - Config object of custom preprocessor.
+// @param config {Object} - Config object of rollupPreprocessor.
+// @param logger {Object} - Karma's logger.
+// @helper helper {Object} - Karma's helper functions.
+function createPreprocessor (args, config, logger, helper)
+{
+    var log = logger.create('preprocessor.rollup');
 
-    let log = logger.create('preprocessor.rollup');
+    config = config || {};
 
     // Don't continue if there is no rollup configuration
     if (config.rollup) {
+        var rollupConfig = config.rollup || {};
+        var bundleConfig = config.bundle || {};
 
-        let options = helper.merge({
-            format: 'es' // default to 'es' format
-        }, config.bundle || {});
-
-        return (content, file, done) => {
-
-            log.debug(' 🗞  Rollup of "%s".', file.originalPath);
+        function preprocess (content, file, done)
+        {
+            log.debug('Processing "%s".', file.originalPath);
 
             try {
-                config.rollup.entry = file.originalPath;
+                rollupConfig.entry = file.originalPath;
 
-                rollup(config.rollup)
-                    .then(bundle => {
+                rollup
+                    .rollup(rollupConfig)
+                    .then(function (bundle)
+                    {
+
                         // Map this file to the dependencies that Rollup just
                         // compiled.
                         dependencies.set(
                             file.originalPath,
-                            bundle.modules.map(b => b.id).filter((op) => op !== file.originalPath));
+                            bundle.modules
+                                .map(function (b) { return b.id })
+                                .filter(function (op) {
+                                    return op !== file.originalPath }))
+
                         // Work backwards from dependencies to see what
                         // relies on this file, then trigger a recompilation of
                         // it.
                         for (let i = 0, list = dependencies.entries(); i < list.length; i += 1) {
-                            let entry = list[i];
-                            let parent = entry[0];
-                            let dependList = entry[1];
-
+                            var entry = list[i];
+                            var parent = entry[0];
+                            var dependList = entry[1]
                             if (dependList.includes(file.originalPath)) {
-                                log.debug(' \n%s depends on \n\t%s\n    Recompiling it.',
+                                log.debug(" \n%s depends on \n\t%s\n    Recompiling it.",
                                     parent, file.originalPath);
                                 changedParents.add(parent);
                                 touchParents();
                             }
                         }
 
-                        let generated = bundle.generate(options);
-                        let processed = generated.code;
+                        if (!bundleConfig.hasOwnProperty('format')) {
+                            bundleConfig.format = 'es';
+                        }
 
-                        if (options.sourceMap === 'inline') {
-                            processed += `# sourceMappingURL=${generated.map.toUrl()}`;
+                        var generated = bundle.generate(bundleConfig);
+                        var processed = generated.code;
+
+                        if (bundleConfig.sourceMap === 'inline') {
+                            var url = generated.map.toUrl();
+                            processed += "\n" + '//# sourceMappingURL=' + url;
                         }
 
                         done(null, processed);
                     })
-                    .catch(error => {
-                        log.error('Failed to process "%s".\n  %s', file.originalPath, error.message);
-                        done(error, null);
+                    .catch(function (error)
+                    {
+                        log.error('%s\n at %s\n%s', e.message, file.originalPath, e.stack);
+                        done(error, null)
                     });
 
-            } catch (e) {
-                log.error('%s\n at %s\n%s', e.message, file.originalPath, e.stack);
-                done(e, null);
             }
-        };
+            catch (exception) {
+                log.error('%s\n at %s', exception.message, file.originalPath);
+                done(exception, null);
+            }
+        }
+
+        return preprocess;
     }
 }
 
 createPreprocessor.$inject = ['args', 'config.rollupPreprocessor', 'logger', 'helper'];
 
-// PUBLISH DI MODULE
-export default {
+module.exports = {
     'preprocessor:rollup': ['factory', createPreprocessor]
 };